Minor optimization for enumerating locked linked lists.

+++ sbbs3/xtrn.cpp
@@ -2,13 +2,13 @@
 
 /* Synchronet external program support routines */
 
-/* $Id: xtrn.cpp,v 1.212 2010/03/13 08:15:08 rswindell Exp $ */
+/* $Id: xtrn.cpp,v 1.215 2011/09/08 02:25:06 deuce Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2010 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2011 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -45,7 +45,7 @@
 	#include <sys/wait.h>	// WEXITSTATUS
 
 	#define TTYDEFCHARS		// needed for ttydefchars definition
-
+	#include <sys/ttydefaults.h>	// Linux - it's motherfucked.
 #if defined(__FreeBSD__)
 	#include <libutil.h>	// forkpty()
 #elif defined(__OpenBSD__) || defined(__NetBSD__) || defined(__DARWIN__)
@@ -386,7 +386,7 @@ int sbbs_t::external(const char* cmdline, long mode, const char* startup_dir)
     PROCESS_INFORMATION process_info;
 	DWORD	hVM;
 	unsigned long	rd;
-    DWORD	wr;
+    unsigned long	wr;
     unsigned long	len;
     DWORD	avail;
 	unsigned long	dummy;
